+++ lib/type_id/uuid.ex
@@ -8,6 +8,11 @@ defmodule TypeID.UUID do
     <<time::big-unsigned-integer-size(48), 7::4, rand_a::12, 2::2, rand_b::62>>
   end
 
+  def uuid4() do
+    <<u0::48, _::4, u1::12, _::2, u2::62>> = :crypto.strong_rand_bytes(16)
+    <<u0::48, 4::4, u1::12, 2::2, u2::62>>
+  end
+
   def binary_to_string(
         <<a1::4, a2::4, a3::4, a4::4, a5::4, a6::4, a7::4, a8::4, b1::4, b2::4, b3::4, b4::4,
           c1::4, c2::4, c3::4, c4::4, d1::4, d2::4, d3::4, d4::4, e1::4, e2::4, e3::4, e4::4,